+++ src/tdastro/rand_nodes/jax_random.py
