+++ tests/tdastro/testing_utils/make_fake_opsim.py
