+++ tests/tdastro/testing_utils/make_opsim_shorten.py
